etl-final/extractor-service/requirements.txt
Django
djangorestframework
kafka-python
requests
pymysql
psycopg2-binary
pandas
openpyxl
